Add support for functional migrations
Functional migrations are not defined by an SQL script but performed
programmatically by providing a function with one argument. The
function receives the connection or database object of the underlying
SQL driver library which is used by the application. The function is
executed within the context of a database transaction.

To reduce the risk which is associated with allowing function calls
via file input, the package which contains the function has to be
defined when initialising a provider (see slot `package`, default is
the `keyword` package which usually doesn't contain functions).

For an example of functional migrations, refer to t/test-suite.lisp
but keep in mind that the test is overly complex as it has to test
multiple drivers.

This change fixes #6.